Rebranded OxRDF to OxSC, and removed links to Vitae RDF.
I'd originally assumed that this skills classification was a simplication of
Vitae's Researcher Development Framework. However, I've since been informed
that it existed before the Vitae RDF, and that the mapping between the two
taxonomies came later. As a result, I've clarified the name to differentiate it
from Vitae's RDF, and improved the rdfs:comment, and replaced the Vitae
conditions of use dcterms:license with CC0.

Vitae have claimed that publishing the mapping between the two categorizations
requires compliance with the Vitae RDF conditions of use:

On 27 March 2013 09:57, John Igoe <john.igoe@vitae.ac.uk> wrote:
> To clarify, from what we can see the Oxford Skills Classification by itself
> (on its own) is not seen as being derived from the Vitae RDF. I’ve attached
> an Excel of our full understanding of the Oxford Skills Classification. It’s
> the mapping from Oxford Skills Classification to the Vitae RDF which is where
> the Vitae RDF conditions of use apply.

These conditions of use¹ state:

> The RDF (ISBN 978-1-906774-18-9), which includes the domains, sub-domains and
> descriptors, and the RDF graphics and Associated Materials may be copied and
> freely reproduced, in any medium, by UK higher education institutions in the
> UK, Research Council institutes and by individual researchers who are members
> of these bodies only, strictly for their own non-commercial training and
> development purposes […]

¹ http://www.vitae.ac.uk/researchers/274251/RDF-conditions-of-use.html

As our data would appear in course *advertising* contexts, we judge that we
would not be able to comply with these conditions of use. As a result, we've
taken the decision to remove the relevant SKOS predicates from the generated
RDF.